cleaning up monthly report gen scripts

+++ tools/reports/makepoints.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 date=$1
 name=points-$date.js
-rm $name
+[ -e $name ] && rm $name
 echo -n 'var points = ' >> $name
 curl -s https://api.waivecar.com/report/$date/points >> $name
 web $name
